+++ public/index.html
@@ -9,7 +9,7 @@
       href="https://fonts.googleapis.com/css2?family=Work+Sans:wght@300;400;500;600;700&display=swap"
       rel="stylesheet"
     />
-    <title>Pausen App</title>
+    <title>PauZen</title>
   </head>
   <body>
     <div id="root"></div>
