+++ src/components/BackButton/BackButton.stories.js
@@ -1,0 +1,8 @@
+import BackButton from './BackButton';
+
+export default {
+  title: 'BackButton',
+  component: BackButton,
+};
+
+export const Default = () => <BackButton />;
