+++ src/components/Card/Card.stories.js
@@ -1,5 +1,7 @@
 import Card from './Card';
 
+const exampleExercise = {};
+
 export default {
   title: 'Card',
   component: Card,
