+++ src/components/CountDownTimer/CountDownTimer.js
@@ -26,6 +26,7 @@ export default function CountDownTimer({exerciseDuration}) {
 
 const StyledParagraph = styled.p`
   font-size: 4rem;
+  line-height: 4rem;
   color: var(--primary-color);
   font-weight: 300;
   min-width: 10.6rem;
