+++ src/components/CountDownTimer/CountDownTimer.stories.js
@@ -1,0 +1,8 @@
+import CountDownTimer from './CountDownTimer';
+
+export default {
+  title: 'CountDownTimer',
+  component: CountDownTimer,
+};
+
+export const Default = () => <CountDownTimer />;
